requirements.txt
websockets==10.4
openpyxl==3.1.1
Pillow==9.4.0
python-docx==0.8.11
python-pptx==0.6.21